backend/requirements.txt
# =============================================================================
# FacPark - Backend Dependencies
# =============================================================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.0  # Required for PyMySQL SSL

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# LLM Clients
google-generativeai>=0.3.0
groq>=0.4.0

# RAG Components
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
nltk>=3.8.0

# Reranker (optional)
# sentence-transformers already includes cross-encoder support

# Vision
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.9.0
pillow>=10.2.0

# Text Processing
unidecode>=1.3.0
symspellpy>=6.7.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.0
python-dateutil>=2.8.0

# Logging & Monitoring
structlog>=24.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
